feat: Add SignUpForm functionality, error handling
This commit will add the SignUpForm functionality including the error
handling.
It will create the SignUpForm event handler via using the useState hook
It will create the handleChange event for filling in the form and
the handleSubmit event to submit the form, via wiring up axios to
connect it to our api.
Error message handling is done and displayed to the user with the
bootstrap alert messages.