feat: Add navbar toggler custom hook
This commit will add a navbar toggler hook, which takes care of the
current state of the navbar icon, whether it is toggled or not.
It creates a reference for the state and if a user clicks outside the
navbar icon, it will update the ref and toggle it.
The eventListener will check on the mouseklick event.
Code is separated into a reusable custom hook.