feat: Add the match create site (initial version)
This commit will add an initial version of the match create site.
It adds the form with all the fields as defined in my backend
as well as the option to upload and change an image.

Styling is not finished yet as well as the functionality. Not
linked to the API yet.