feat: Add initial version of Match page
This commit will add an initial version of the Match page functionality
which will then be rendered on teh MatchPage

The API call is executed for an existing match and it's values are
rendered, even though not finished yet.

Furthermore, a helper component is created in a dedicated folder.
It does translate the date object, which is returned fromt he API,
into a date format, which is desired. It transforms it in a way
that we have it like 13. June 2023, so it is also adding a dot after
the day.
This component is generic but gets now used in the match page.