feat: Add context to match page
This commit will add further context to the Match page. It will include
the rest of the missing fields from the DB.
It will also add the Attending icon as well as the comments icon.

The Attending functionality is wired up so that a logged in user,
including the owner of a match post, can attend a match by clicking
on the icon. The icon will be persistend tred and the count will
increase by one.
The unattend logic is just the reverse logic of the above.
If a user is not logged in and wants to attend a match, the icon will
turn black and an overlay text gets displayed to log in first.