feat: Add logic to the Matches & MySchedule Page
This commit will add the logic to the Matches and the MySchedule page
to display only the relevant data, with the filter defined in the
App.js.
It will display either all matches or in the MySchedule page, only
the ones the logged in user is attending.
Re-fetch happens when the URL change happens - a user clicks between
the mySchedule and the Matches site.