feat: Add search feature and update match_date
This commit will add a search bar feature to the matches and the
mySchedule page. It will create a smart search, adding a delay of 1s
before the api request is sent, in order to have more efficient search.

It will also adjust the match_date reformating custom hook. Since the
DRF part got updated and we receive now a match_date: 13 Aug 2023
format, the original FormatDate hook is no longer required. A new
custom hook called FormatDay is created instead, in order to add a
dot behind the day, like 13. Aug 2023