feat: Add infinite scroll feature
The API limits the results paginated to 5 results. This feature
will make sure, we have an autoscroll feature and loading more and
more elements.
However, the infinite scroll is done in the way that it checks if the
accumulated result id is not the same as the current result id and
therefore overcomes the "duplicate" issue.