feat: Add gun create from to project
This commit will add the MyGuns app to the project. A new add gun
button has been created in the navbar.
A new gun create form has been added with styling like the add match
site.
The form fields are created as well as the logic to upload and
change an image.