feat: Add functionality to add gun to database
This commit will add the functionality to add a gun to the myGuns
database. It will import axios and create an api call onSubmit.
Furthermore, the functionality of cancelling the form fill in is added,
which brings the user back to the last site he was on.
Bootstrap alert messages will be shown, if the mandatory fields are not
filled in, however, the alert functionality is added to all fields.