feat: Add GunsPage to app and display myGuns list
This commit will add the logic that if a user is logged in and
clicks ot the MyGuns url, he will see all guns he has added to the
database.
He will only see his guns and only while logged in.
A delay timeout is set that it waits 1s before rendering, this feature
will be used once the search field is added.
An error state is created to make sure, a logged out user trying to
access /guns will see a 401 error, that he is not authorized to view
the content. This is a important feature because no one except the
owner should be able to see the collection database.

Styling is not final yet, there are still placeholders, which will
be deleted later.