feat: Add search feature to myGuns site
This commit will add the search feature to the myGuns site. The user
is able to search after brand, model or serial number.
If the user is logged out and tries to access the /guns url, he
will be able to see the search bar, together with the 401 no permission
error, but he will not be able to search. This is an important
security feature.