feat: Add edit functionality for match and gun
This commit will add the edit pages for gun and match. It will
give the possibility to update all fields, including the image.
The match_date field needs some special treatment, since we
currently get the match_date in the form of 12 Aug 2023 from the
API. This has been adjusted in the DRF part because I want that the
match_date field looks like that.
However, in order to be complaint with the date form field, I had
to create the formatMatchDate helper function. It will correctly
adjust hte field.
After this commit it is also possible again to create a new match,
since this was no longer possible due to earlier date adjustment on
the DRF part.