feat: Add match comments functionality
This commit will add the match comments functionality to the project.
The create comment form is added and styling is done to reflect the
overall styling.
Once a comment is submitted, it gets correctly reflected by the
comments count icon on the matches page, however, the comment is not
visible yet.