feat: Add delete comment option & update style
This commit will add the delete comment option to the comments function.
It will check if the current user is the owner of the commend and if so,
it will show the edit icon and let him delete the comment.
It will also decrease the comments count by 1, if the comment got
deleted.
The commit will further adjust the styling of the comment section as
well as change the icon for the "edit" functionality for all edit
option in the project (match, gun, comment)