feat: Add popular matches mobile & styling
This commit will add a dedicated version of the popular matches
component for the mobile view. Only the top 3 will be displayed,
where as for the desktop, the top 6 will be shown.
It adds styling to the content, including a link to the related
match, so a user can directly click on it and see details / attend.