fix: Adjust state handling of AttendingShooters
This commit will fix two problems with the AttendingShooters component.
1. It will add an additional ternary when rendering the attendings
in the AttendingShooters component to make sure, if no one is
attending a match, the spinner will not run endlessly and rather a
message gets displayed telling that no one is currently attending.
This will also solve the same problem, if a new match is created,
because at that point in time, no one is attending.

2. The attending count is passed as a prop to the AttendingShooters
component and added as a dependency to the useEffect hoook inside it.
Therefore, every time there is a change in the atteindings count,
the AttendingShooters component gets re-rendered. This is important
if the user is on the MatchPage, the detail page displaying the
individual match with the comments and the AttendingShooters component.
If he clicks to attend the match, the component must be immediately
updated and vice versae.