feat: Add profile page as first draft
This commit creates the profile page, which will dispaly all data
of the particular shooter.
Styling and data is raw for now, will be enhanced in later commit.
It does also include the matches, which the respective user has
created. If he did not create a match yet, the info will be shown.