feat: Add filter for MySchedule
This commit will add a client-side filter for the MySchedule site.
It will display only events, which are in the future and the sorting
is done to display the most recent first.
This behavior is different than in the Matches site, since the
matches site will also display past events as kind of history, and
sort based on the created_on date.