feat: Add useRedirect hook
This commit will add the custom useRedirect hook to the project.
It will make sure, the user gets redirected to the main page, if he
is trying to access content, which he should not or is not allowed.
Furthermore, it will simplify the logic when created a user account
and then logging in.