feat: Add popularMatches to gun detail
This commit will add the popularMatches component to the gun detail
view for desktop and mobile. It will again encourage the user
to also attend a popular match.
It will also implement the FormatDay function for the created_/
updated_at fields. With this, we have a uniform styling of the date
thorughout the app.