feat: Add error handling modal
This commit will add a error handling modal to the application.
There are several try / catch blocks in the app. Logging errors to the
console is not a good idea for production. However, just having
the user in "unknown" state, not knowing that there was an issue is
also not a good idea.
Therefore a error modal has been created and called the catch block.
For testing purpose, only on the GunPage app, will be added to the
others later.