fix: Correct url form field
This commit will fix an error of the profile edit social_media
form field. It was not defined as URL, where as the DRF back end did
expect an URLfield.
Type got updated in the form and therefore client side verification
is done.
Furhtermore a placeholder text is added, so the user knows exactly
which format to enter.
The placeholder text also got added to the mailaddress.