fix: Correct controll id on form
This commit will fix a problem in the SignIn and SignUp form.
Since I have added in an earlier step the htmlFor attribute and an
in in each form field, this a duplicate with the already set
controlID set. React Bootstrap will dynamically create the labels
for this when set.