refactor: Add updated version of GunsPage
This commit will refactor the GunsPage. It will remove the option
to display an error of not being logged in, if an unauthenticated user
tries to access the url without being logged in. I somehow do not like
it, therefore I have implemented that the user gets pushed to the main
page, which is accessible for everyone.